refactor: move AccessControlDefault to accesscontrol_default.hpp
