+++ include/open62541pp/open62541pp.hpp
@@ -11,14 +11,8 @@
 #include "open62541pp/monitoreditem.hpp"
 #include "open62541pp/node.hpp"
 #include "open62541pp/nodeids.hpp"
-#include "open62541pp/plugin/accesscontrol.hpp"
-#include "open62541pp/plugin/create_certificate.hpp"
-#include "open62541pp/plugin/log.hpp"
-#include "open62541pp/plugin/nodestore.hpp"
-#include "open62541pp/plugin/pluginadapter.hpp"
 #include "open62541pp/result.hpp"
 #include "open62541pp/server.hpp"
-#include "open62541pp/services/services.hpp"
 #include "open62541pp/session.hpp"
 #include "open62541pp/span.hpp"
 #include "open62541pp/subscription.hpp"
@@ -29,3 +23,12 @@
 #include "open62541pp/types_composed.hpp"
 #include "open62541pp/typewrapper.hpp"
 #include "open62541pp/wrapper.hpp"
+
+#include "open62541pp/plugin/accesscontrol.hpp"
+#include "open62541pp/plugin/accesscontrol_default.hpp"
+#include "open62541pp/plugin/create_certificate.hpp"
+#include "open62541pp/plugin/log.hpp"
+#include "open62541pp/plugin/nodestore.hpp"
+#include "open62541pp/plugin/pluginadapter.hpp"
+
+#include "open62541pp/services/services.hpp"
