+++ src/client.cpp
@@ -14,7 +14,7 @@
 #include "open62541pp/detail/open62541/common.h"
 #include "open62541pp/exception.hpp"
 #include "open62541pp/node.hpp"
-#include "open62541pp/plugin/accesscontrol.hpp"  // Login
+#include "open62541pp/plugin/accesscontrol_default.hpp"  // Login
 #include "open62541pp/result.hpp"
 #include "open62541pp/services/attribute_highlevel.hpp"  // readValue
 #include "open62541pp/services/subscription.hpp"
